Centralize getCategoryVisibility for ALL flagCards #42
Before this the querySelector only checked the first flag but also ran the function for all flags individually.
Now it only runs onces for all flags.
